@@ -2,17 +2,21 @@ package com.androiddevs.mvvmnewsapp
 
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
-import androidx.lifecycle.ViewModelProvider
+import androidx.databinding.DataBindingUtil
+import androidx.navigation.fragment.NavHostFragment
 import androidx.navigation.fragment.findNavController
 import androidx.navigation.ui.setupWithNavController
-import com.androiddevs.mvvmnewsapp.commonviewmodel.NewsViewModel
-import kotlinx.android.synthetic.main.activity_main.*
+import com.androiddevs.mvvmnewsapp.databinding.ActivityMainBinding
 
 class NewsActivity : AppCompatActivity() {
 
+    private lateinit var binding: ActivityMainBinding
+
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_main)
-        bottomNavigationView.setupWithNavController(newsNavHostFragment.findNavController())
+        binding = DataBindingUtil.setContentView(this, R.layout.activity_main)
+        val navHostFragment =
+            supportFragmentManager.findFragmentById(R.id.newsNavHostFragment) as NavHostFragment
+        binding.bottomNavigationView.setupWithNavController(navHostFragment.findNavController())
     }
 }
