@@ -1,34 +1,47 @@
 package com.androiddevs.mvvmnewsapp.articles
 
 import android.os.Bundle
+import android.view.LayoutInflater
 import android.view.View
+import android.view.ViewGroup
 import android.webkit.WebViewClient
 import androidx.fragment.app.Fragment
 import androidx.lifecycle.ViewModelProvider
 import androidx.navigation.fragment.navArgs
-import com.androiddevs.mvvmnewsapp.R
 import com.androiddevs.mvvmnewsapp.commonviewmodel.NewsViewModel
+import com.androiddevs.mvvmnewsapp.databinding.FragmentArticleBinding
 import com.google.android.material.snackbar.Snackbar
-import kotlinx.android.synthetic.main.fragment_article.*
 
-class ArticleFragment : Fragment(R.layout.fragment_article){
+class ArticleFragment : Fragment(){
     private val viewModel by lazy {
         ViewModelProvider(requireActivity()).get(NewsViewModel::class.java)
     }
     private val args: ArticleFragmentArgs by navArgs()
 
+    private lateinit var binding: FragmentArticleBinding
+    override fun onCreateView(
+        inflater: LayoutInflater,
+        container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View? {
+        binding = FragmentArticleBinding.inflate(inflater, container, false)
+        return binding.root
+    }
+
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
         val article = args.article
 
-        webView.apply {
-            webViewClient = WebViewClient()
-            loadUrl(article.url)
-        }
+        binding.apply {
+            webView.apply {
+                webViewClient = WebViewClient()
+                loadUrl(article.url)
+            }
 
-        fab.setOnClickListener {
-            viewModel.saveArticle(article)
-            Snackbar.make(view, "Article saved successfully", Snackbar.LENGTH_SHORT).show()
+            fab.setOnClickListener {
+                viewModel.saveArticle(article)
+                Snackbar.make(view, "Article saved successfully", Snackbar.LENGTH_SHORT).show()
+            }
         }
     }
 }
