@@ -2,7 +2,9 @@ package com.androiddevs.mvvmnewsapp.breakingNews
 
 import android.os.Bundle
 import android.util.Log
+import android.view.LayoutInflater
 import android.view.View
+import android.view.ViewGroup
 import android.widget.AbsListView
 import android.widget.Toast
 import androidx.fragment.app.Fragment
@@ -14,9 +16,9 @@ import androidx.recyclerview.widget.RecyclerView
 import com.androiddevs.mvvmnewsapp.R
 import com.androiddevs.mvvmnewsapp.commonviewmodel.NewsViewModel
 import com.androiddevs.mvvmnewsapp.commonAdapters.NewsAdapter
+import com.androiddevs.mvvmnewsapp.databinding.FragmentBreakingNewsBinding
 import com.androiddevs.mvvmnewsapp.util.Constants.Companion.QUERY_PAGE_SIZE
 import com.androiddevs.mvvmnewsapp.util.Resource
-import kotlinx.android.synthetic.main.fragment_breaking_news.*
 
 class BreakingNewsFragment : Fragment(R.layout.fragment_breaking_news){
     private val TAG = "BreakingNewsFragment"
@@ -24,6 +26,16 @@ class BreakingNewsFragment : Fragment(R.layout.fragment_breaking_news){
         ViewModelProvider(requireActivity()).get(NewsViewModel::class.java)
     }
     private lateinit var newsAdapter: NewsAdapter
+
+    private lateinit var binding: FragmentBreakingNewsBinding
+    override fun onCreateView(
+        inflater: LayoutInflater,
+        container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View? {
+        binding = FragmentBreakingNewsBinding.inflate(inflater, container, false)
+        return binding.root
+    }
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
@@ -67,20 +79,19 @@ class BreakingNewsFragment : Fragment(R.layout.fragment_breaking_news){
 
     private fun setupRecyclerView() {
         newsAdapter = NewsAdapter()
-        rvBreakingNews.apply {
-            adapter = newsAdapter
-            layoutManager = LinearLayoutManager(requireActivity())
-            addOnScrollListener(this@BreakingNewsFragment.scrollListener)
+        binding.apply {
+            rvBreakingNews.apply {
+                adapter = newsAdapter
+                layoutManager = LinearLayoutManager(requireActivity())
+                addOnScrollListener(this@BreakingNewsFragment.scrollListener)
+            }
         }
     }
 
     private fun setOnNewsItemClick() {
         newsAdapter.setOnItemClickListener {
-            val bundle = Bundle().apply {
-                putSerializable("article", it)
-            }
             findNavController().navigate(
-                R.id.action_breakingNewsFragment_to_articleFragment, bundle
+                BreakingNewsFragmentDirections.actionBreakingNewsFragmentToArticleFragment(it)
             )
         }
     }
@@ -96,7 +107,7 @@ class BreakingNewsFragment : Fragment(R.layout.fragment_breaking_news){
                         val totalPages = newsResponse.totalResults / QUERY_PAGE_SIZE + 2
                         isLastPage = viewModel.breakingNewsPage == totalPages
                         if (isLastPage) {
-                            rvBreakingNews.setPadding(0, 0, 0, 0)
+                            binding.rvBreakingNews.setPadding(0, 0, 0, 0)
                         }
                     }
                     hideProgressBar()
@@ -113,12 +124,12 @@ class BreakingNewsFragment : Fragment(R.layout.fragment_breaking_news){
     }
 
     private fun hideProgressBar() {
-        paginationProgressBar.visibility = View.INVISIBLE
+        binding.paginationProgressBar.visibility = View.INVISIBLE
         isLoading = false
     }
 
     private fun showProgressBar() {
-        paginationProgressBar.visibility = View.VISIBLE
+        binding.paginationProgressBar.visibility = View.VISIBLE
         isLoading = true
     }
 }
